I pulled Damians branch and manually added some of the remaining stuff.
There is a settings view that shows settings based off employee, manager, and admin.
I added admin to the database and created an admin account that reviews all employees and managers, eventually may only oversee those in a certain department/branch etc.
There is a clock in/out view and a select hours view that is in the settings under admin that saves the decision to the database.
I am working on the leave requests and will add it once i fix it to the manager dash and database.
will add the admin dash in the next commit